tests/requirements-test.txt
pytest
pytest-asyncio
pytest-timeout
pytest-playwright
playwright
httpx
psycopg2-binary
redis
tenacity
